+++ src/index.html
@@ -31,7 +31,10 @@
     <div id="root"></div>
                        
     <script type="module" src="/src/pages/feed/feed.js"></script>
+    <script type="module" src="/src/pages/login/login.js"></script>
+    <script type="module" src="pages/subscribe/subscribe.js"></script>
     <script type="module" src="router.js"></script>
+
     <!-- The core Firebase JS SDK is always required and must be listed first -->
     <script src="https://www.gstatic.com/firebasejs/8.2.1/firebase-app.js"></script>
